requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

langchain==0.1.16
langgraph==0.0.40
langchain-community==0.0.32

groq==0.5.0

pinecone-client==3.2.2

sentence-transformers==2.6.1
faiss-cpu==1.7.4

pymupdf==1.24.0

networkx==3.2.1

requests==2.31.0
python-dotenv==1.0.1

pydantic==2.6.4

tqdm==4.66.2

arxiv==2.1.0
aiohttp==3.9.3
graphviz==0.20.1
